Convert to/from an adjacency matrix.

There are two ways that peptide/protein matches are commonly stored: either as a vector or an adjacency matrix. The functions described below convert between these two format.

makeAdjacencyMatrix(
  x,
  split = ";",
  peptide = psmVariables(x)["peptide"],
  protein = psmVariables(x)["protein"]
)

makePeptideProteinVector(m, collapse = ";")

Arguments

x

Either an instance of class PSM or a character. See example below for details.

split

character(1) defining how to split the string of protein identifiers (using strsplit()). Default is ";". If NULL, splitting is ignored.

peptide

character(1) indicating the name of the variable that defines peptides in the PSM object. Default is the peptide PSM variable as defined in psmVariables().

protein

character(1) indicating the name of the variable that defines proteins in the PSM object. Default is the peptide PSM variable as defined in psmVariables().Default is DatanbaseAccess.

m

An adjacency matrix.

collapse

character(1) indicating how to collapse protein names for non-proteotypic peptides. Default is ";".

Value

A peptide-by-protein adjacency matrix or peptide/protein vector.

Details

The makeAdjacencyMatrix() function creates a peptide-by-protein adjacency matrix from a character or an instance of class PSM().

The character is formatted as x <- c("ProtA", "ProtB", "ProtA;ProtB", ...), as commonly encoutered in proteomics data spreadsheets. It defines that the first peptide is mapped to protein "ProtA", the second one to protein "ProtB", the third one to "ProtA" and "ProtB", and so on. The resulting matrix contain length(x) rows an as many columns as there are unique protein idenifiers in x. The column are always named after the protein idenifiers. If the peptide/protein vector is named and the names are unique, these are then used to name to matrix rows.

The makePeptideProteinVector() function does the opposite operation, taking an adjacency matrix as input and retruning a peptide/protein vector. The matrix colnames are used to populate the vector and the matrix rownames are used to name the vector elements.

Author

Laurent Gatto

Examples

## -----------------------
## From a character
## -----------------------

## Protein vector without names
prots <- c("ProtA", "ProtB", "ProtA;ProtB")
makeAdjacencyMatrix(prots)
#>      ProtA ProtB
#> [1,]     1     0
#> [2,]     0     1
#> [3,]     1     1

## Named protein vector
names(prots) <- c("pep1", "pep2", "pep3")
prots
#>          pep1          pep2          pep3 
#>       "ProtA"       "ProtB" "ProtA;ProtB" 
m <- makeAdjacencyMatrix(prots)
m
#>      ProtA ProtB
#> pep1     1     0
#> pep2     0     1
#> pep3     1     1

## Back to vector
vec <- makePeptideProteinVector(m)
vec
#>          pep1          pep2          pep3 
#>       "ProtA"       "ProtB" "ProtA;ProtB" 
identical(prots, vec)
#> [1] TRUE

## -----------------------
## From a PSM object
## -----------------------

f <- msdata::ident(full.names = TRUE, pattern = "TMT")
psm <- PSM(f) |>
       filterPsmDecoy() |>
       filterPsmRank()
#> Removed 2896 decoy hits.
#> Removed 155 PSMs with rank > 1.
psm
#> PSM with 2751 rows and 35 columns.
#> Spectra: 2685 unique
#>   db: 2751 target, 0 decoy
#>   ranks: 1:2751 
#> Peptides: 2324 unique, 33 multiple
#> Proteins: 1504 
#> names(35): sequence spectrumID ... subReplacementResidue subLocation
adj <- makeAdjacencyMatrix(psm)
dim(adj)
#> [1] 2357 1504
adj[1:10, 1:4]
#>                                    colnames
#> rownames                            ECA2006 ECA1676 ECA3009 ECA1420
#>   RQCRTDFLNYLR                            1       0       0       0
#>   ESVALADQVTCVDWRNRKATKK                  0       1       0       0
#>   QRMARTSDKQQSIRFLERLCGR                  0       0       1       0
#>   DGGPAIYGHERVGRNAKKFKCLKFR               0       0       0       1
#>   CIDRARHVEVQIFGDGKGRVVALGERDCSLQRR       0       0       0       0
#>   VGRCRPIINYLASPGGTSER                    0       0       0       0
#>   QRLDEHCVGVGQNALLLGR                     0       0       0       0
#>   VDYQGKKVVIIGLGLTGLSCVDFFLARGVVPR        0       0       0       0
#>   TLIIERCR                                0       0       0       0
#>   QEVESCTTIEKIWVIDLKIK                    0       0       0       0
## Peptides with rowSums > 1 match multiple proteins.
## Use filterPsmNonProteotypic() to filter these out.
table(rowSums(adj))
#> 
#>    1    2    3    4 
#> 2324   28    3    2