Calculate ions produced by fragmentation with variable modifications

Calculate ions produced by fragmentation with variable modifications

calculateFragments2(
  sequence,
  type = c("b", "y"),
  z = 1,
  fixed_modifications = c(C = 57.02146),
  variable_modifications = numeric(),
  max_mods = Inf,
  neutralLoss = defaultNeutralLoss(),
  verbose = TRUE,
  modifications = NULL
)

Arguments

sequence

character() providing a peptide sequence.

type

character vector of target ions; possible values: c("a", "b", "c", "x", "y", "z"). Default is type = c("b", "y").

z

numeric with a desired charge state; default is 1.

fixed_modifications

A named numeric vector of used fixed modifications. The name must correspond to the one-letter-code of the modified amino acid and the numeric value must represent the mass that should be added to the original amino accid mass, default: Carbamidomethyl modifications = c(C = 57.02146). Use Nterm or Cterm as names for modifications that should be added to the amino respectively carboxyl-terminus.

variable_modifications

A named numeric vector of variable modifications. Depending on the maximum number of modifications (max_mods), all possible combinations are returned.

max_mods

A numeric indicating the maximum number of variable modifications allowed on the sequence at once. Does not include fixed modifications. Default value is positive infinity.

neutralLoss

list, it has to have two named elments, namely water and ammonia that contain a character vector which type of neutral loss should be calculated. Currently neutral loss on the C terminal "Cterm", at the amino acids c("D", "E", "S", "T") for "water" (shown with an ⁠_⁠) an d c("K", "N", "Q", "R") for "ammonia" (shown with an *) are supported.

verbose

logical(1). If TRUE (default) the used modifications are printed.

modifications

Named numeric(). Deprecated modifications parameter. Will override fixed_modifications but is set to NULL by default. Please refrain from using it, opt for fixed_modifications instead.

Value

A dataframe showing all the ions produced by fragmentation with all possible combinations of modifications. The used variable modifications are displayed in the peptide column through the use of amino acids within brackets. Fixed modifications are not displayed.

Examples

calculateFragments2(sequence = "ARGSHKATC", type = c("b", "y"), z = 1, 
fixed_modifications = c(C = 57), variable_modifications = c(S = 79, Y = 79, T = 79),
max_mods = 2)
#> Fixed modifications used: C=57
#> Variable modifications used: S=79, Y=79, T=79
#>             mz ion type pos z      seq           peptide
#> 1     72.04439  b1    b   1 1        A         ARGSHKATC
#> 2    228.14550  b2    b   2 1       AR         ARGSHKATC
#> 3    285.16696  b3    b   3 1      ARG         ARGSHKATC
#> 4    372.19899  b4    b   4 1     ARGS         ARGSHKATC
#> 5    509.25790  b5    b   5 1    ARGSH         ARGSHKATC
#> 6    637.35286  b6    b   6 1   ARGSHK         ARGSHKATC
#> 7    708.38997  b7    b   7 1  ARGSHKA         ARGSHKATC
#> 8    809.43765  b8    b   8 1 ARGSHKAT         ARGSHKATC
#> 9    179.02703  y1    y   1 1        C         ARGSHKATC
#> 10   280.07471  y2    y   2 1       TC         ARGSHKATC
#> 11   351.11182  y3    y   3 1      ATC         ARGSHKATC
#> 12   479.20678  y4    y   4 1     KATC         ARGSHKATC
#> 13   616.26569  y5    y   5 1    HKATC         ARGSHKATC
#> 14   703.29772  y6    y   6 1   SHKATC         ARGSHKATC
#> 15   760.31918  y7    y   7 1  GSHKATC         ARGSHKATC
#> 16   916.42029  y8    y   8 1 RGSHKATC         ARGSHKATC
#> 17   491.24733 b5_   b_   5 1    ARGSH         ARGSHKATC
#> 18   619.34229 b6_   b_   6 1   ARGSHK         ARGSHKATC
#> 19   690.37940 b7_   b_   7 1  ARGSHKA         ARGSHKATC
#> 20   791.42708 b8_   b_   8 1 ARGSHKAT         ARGSHKATC
#> 21   333.10126 y3_   y_   3 1      ATC         ARGSHKATC
#> 22   461.19622 y4_   y_   4 1     KATC         ARGSHKATC
#> 23   598.25513 y5_   y_   5 1    HKATC         ARGSHKATC
#> 24   685.28716 y6_   y_   6 1   SHKATC         ARGSHKATC
#> 25   742.30862 y7_   y_   7 1  GSHKATC         ARGSHKATC
#> 26   898.40973 y8_   y_   8 1 RGSHKATC         ARGSHKATC
#> 27   161.01647 y1_   y_   1 1        C         ARGSHKATC
#> 28   262.06415 y2_   y_   2 1       TC         ARGSHKATC
#> 29   268.14041 b3*   b*   3 1      ARG         ARGSHKATC
#> 30   355.17244 b4*   b*   4 1     ARGS         ARGSHKATC
#> 31   492.23135 b5*   b*   5 1    ARGSH         ARGSHKATC
#> 32   620.32631 b6*   b*   6 1   ARGSHK         ARGSHKATC
#> 33   691.36342 b7*   b*   7 1  ARGSHKA         ARGSHKATC
#> 34   792.41110 b8*   b*   8 1 ARGSHKAT         ARGSHKATC
#> 35   462.18023 y4*   y*   4 1     KATC         ARGSHKATC
#> 36   599.23914 y5*   y*   5 1    HKATC         ARGSHKATC
#> 37   686.27117 y6*   y*   6 1   SHKATC         ARGSHKATC
#> 38   743.29263 y7*   y*   7 1  GSHKATC         ARGSHKATC
#> 39   899.39374 y8*   y*   8 1 RGSHKATC         ARGSHKATC
#> 40    72.04439  b1    b   1 1        A     ARGS[79]HKATC
#> 41   228.14550  b2    b   2 1       AR     ARGS[79]HKATC
#> 42   285.16696  b3    b   3 1      ARG     ARGS[79]HKATC
#> 43   451.19899  b4    b   4 1     ARGS     ARGS[79]HKATC
#> 44   588.25790  b5    b   5 1    ARGSH     ARGS[79]HKATC
#> 45   716.35286  b6    b   6 1   ARGSHK     ARGS[79]HKATC
#> 46   787.38997  b7    b   7 1  ARGSHKA     ARGS[79]HKATC
#> 47   888.43765  b8    b   8 1 ARGSHKAT     ARGS[79]HKATC
#> 48   179.02703  y1    y   1 1        C     ARGS[79]HKATC
#> 49   280.07471  y2    y   2 1       TC     ARGS[79]HKATC
#> 50   351.11182  y3    y   3 1      ATC     ARGS[79]HKATC
#> 51   479.20678  y4    y   4 1     KATC     ARGS[79]HKATC
#> 52   616.26569  y5    y   5 1    HKATC     ARGS[79]HKATC
#> 53   782.29772  y6    y   6 1   SHKATC     ARGS[79]HKATC
#> 54   839.31918  y7    y   7 1  GSHKATC     ARGS[79]HKATC
#> 55   995.42029  y8    y   8 1 RGSHKATC     ARGS[79]HKATC
#> 56   570.24733 b5_   b_   5 1    ARGSH     ARGS[79]HKATC
#> 57   698.34229 b6_   b_   6 1   ARGSHK     ARGS[79]HKATC
#> 58   769.37940 b7_   b_   7 1  ARGSHKA     ARGS[79]HKATC
#> 59   870.42708 b8_   b_   8 1 ARGSHKAT     ARGS[79]HKATC
#> 60   333.10126 y3_   y_   3 1      ATC     ARGS[79]HKATC
#> 61   461.19622 y4_   y_   4 1     KATC     ARGS[79]HKATC
#> 62   598.25513 y5_   y_   5 1    HKATC     ARGS[79]HKATC
#> 63   764.28716 y6_   y_   6 1   SHKATC     ARGS[79]HKATC
#> 64   821.30862 y7_   y_   7 1  GSHKATC     ARGS[79]HKATC
#> 65   977.40973 y8_   y_   8 1 RGSHKATC     ARGS[79]HKATC
#> 66   161.01647 y1_   y_   1 1        C     ARGS[79]HKATC
#> 67   262.06415 y2_   y_   2 1       TC     ARGS[79]HKATC
#> 68   268.14041 b3*   b*   3 1      ARG     ARGS[79]HKATC
#> 69   434.17244 b4*   b*   4 1     ARGS     ARGS[79]HKATC
#> 70   571.23135 b5*   b*   5 1    ARGSH     ARGS[79]HKATC
#> 71   699.32631 b6*   b*   6 1   ARGSHK     ARGS[79]HKATC
#> 72   770.36342 b7*   b*   7 1  ARGSHKA     ARGS[79]HKATC
#> 73   871.41110 b8*   b*   8 1 ARGSHKAT     ARGS[79]HKATC
#> 74   462.18023 y4*   y*   4 1     KATC     ARGS[79]HKATC
#> 75   599.23914 y5*   y*   5 1    HKATC     ARGS[79]HKATC
#> 76   765.27117 y6*   y*   6 1   SHKATC     ARGS[79]HKATC
#> 77   822.29263 y7*   y*   7 1  GSHKATC     ARGS[79]HKATC
#> 78   978.39374 y8*   y*   8 1 RGSHKATC     ARGS[79]HKATC
#> 79    72.04439  b1    b   1 1        A     ARGSHKAT[79]C
#> 80   228.14550  b2    b   2 1       AR     ARGSHKAT[79]C
#> 81   285.16696  b3    b   3 1      ARG     ARGSHKAT[79]C
#> 82   372.19899  b4    b   4 1     ARGS     ARGSHKAT[79]C
#> 83   509.25790  b5    b   5 1    ARGSH     ARGSHKAT[79]C
#> 84   637.35286  b6    b   6 1   ARGSHK     ARGSHKAT[79]C
#> 85   708.38997  b7    b   7 1  ARGSHKA     ARGSHKAT[79]C
#> 86   888.43765  b8    b   8 1 ARGSHKAT     ARGSHKAT[79]C
#> 87   179.02703  y1    y   1 1        C     ARGSHKAT[79]C
#> 88   359.07471  y2    y   2 1       TC     ARGSHKAT[79]C
#> 89   430.11182  y3    y   3 1      ATC     ARGSHKAT[79]C
#> 90   558.20678  y4    y   4 1     KATC     ARGSHKAT[79]C
#> 91   695.26569  y5    y   5 1    HKATC     ARGSHKAT[79]C
#> 92   782.29772  y6    y   6 1   SHKATC     ARGSHKAT[79]C
#> 93   839.31918  y7    y   7 1  GSHKATC     ARGSHKAT[79]C
#> 94   995.42029  y8    y   8 1 RGSHKATC     ARGSHKAT[79]C
#> 95   491.24733 b5_   b_   5 1    ARGSH     ARGSHKAT[79]C
#> 96   619.34229 b6_   b_   6 1   ARGSHK     ARGSHKAT[79]C
#> 97   690.37940 b7_   b_   7 1  ARGSHKA     ARGSHKAT[79]C
#> 98   870.42708 b8_   b_   8 1 ARGSHKAT     ARGSHKAT[79]C
#> 99   412.10126 y3_   y_   3 1      ATC     ARGSHKAT[79]C
#> 100  540.19622 y4_   y_   4 1     KATC     ARGSHKAT[79]C
#> 101  677.25513 y5_   y_   5 1    HKATC     ARGSHKAT[79]C
#> 102  764.28716 y6_   y_   6 1   SHKATC     ARGSHKAT[79]C
#> 103  821.30862 y7_   y_   7 1  GSHKATC     ARGSHKAT[79]C
#> 104  977.40973 y8_   y_   8 1 RGSHKATC     ARGSHKAT[79]C
#> 105  161.01647 y1_   y_   1 1        C     ARGSHKAT[79]C
#> 106  341.06415 y2_   y_   2 1       TC     ARGSHKAT[79]C
#> 107  268.14041 b3*   b*   3 1      ARG     ARGSHKAT[79]C
#> 108  355.17244 b4*   b*   4 1     ARGS     ARGSHKAT[79]C
#> 109  492.23135 b5*   b*   5 1    ARGSH     ARGSHKAT[79]C
#> 110  620.32631 b6*   b*   6 1   ARGSHK     ARGSHKAT[79]C
#> 111  691.36342 b7*   b*   7 1  ARGSHKA     ARGSHKAT[79]C
#> 112  871.41110 b8*   b*   8 1 ARGSHKAT     ARGSHKAT[79]C
#> 113  541.18023 y4*   y*   4 1     KATC     ARGSHKAT[79]C
#> 114  678.23914 y5*   y*   5 1    HKATC     ARGSHKAT[79]C
#> 115  765.27117 y6*   y*   6 1   SHKATC     ARGSHKAT[79]C
#> 116  822.29263 y7*   y*   7 1  GSHKATC     ARGSHKAT[79]C
#> 117  978.39374 y8*   y*   8 1 RGSHKATC     ARGSHKAT[79]C
#> 118   72.04439  b1    b   1 1        A ARGS[79]HKAT[79]C
#> 119  228.14550  b2    b   2 1       AR ARGS[79]HKAT[79]C
#> 120  285.16696  b3    b   3 1      ARG ARGS[79]HKAT[79]C
#> 121  451.19899  b4    b   4 1     ARGS ARGS[79]HKAT[79]C
#> 122  588.25790  b5    b   5 1    ARGSH ARGS[79]HKAT[79]C
#> 123  716.35286  b6    b   6 1   ARGSHK ARGS[79]HKAT[79]C
#> 124  787.38997  b7    b   7 1  ARGSHKA ARGS[79]HKAT[79]C
#> 125  967.43765  b8    b   8 1 ARGSHKAT ARGS[79]HKAT[79]C
#> 126  179.02703  y1    y   1 1        C ARGS[79]HKAT[79]C
#> 127  359.07471  y2    y   2 1       TC ARGS[79]HKAT[79]C
#> 128  430.11182  y3    y   3 1      ATC ARGS[79]HKAT[79]C
#> 129  558.20678  y4    y   4 1     KATC ARGS[79]HKAT[79]C
#> 130  695.26569  y5    y   5 1    HKATC ARGS[79]HKAT[79]C
#> 131  861.29772  y6    y   6 1   SHKATC ARGS[79]HKAT[79]C
#> 132  918.31918  y7    y   7 1  GSHKATC ARGS[79]HKAT[79]C
#> 133 1074.42029  y8    y   8 1 RGSHKATC ARGS[79]HKAT[79]C
#> 134  570.24733 b5_   b_   5 1    ARGSH ARGS[79]HKAT[79]C
#> 135  698.34229 b6_   b_   6 1   ARGSHK ARGS[79]HKAT[79]C
#> 136  769.37940 b7_   b_   7 1  ARGSHKA ARGS[79]HKAT[79]C
#> 137  949.42708 b8_   b_   8 1 ARGSHKAT ARGS[79]HKAT[79]C
#> 138  412.10126 y3_   y_   3 1      ATC ARGS[79]HKAT[79]C
#> 139  540.19622 y4_   y_   4 1     KATC ARGS[79]HKAT[79]C
#> 140  677.25513 y5_   y_   5 1    HKATC ARGS[79]HKAT[79]C
#> 141  843.28716 y6_   y_   6 1   SHKATC ARGS[79]HKAT[79]C
#> 142  900.30862 y7_   y_   7 1  GSHKATC ARGS[79]HKAT[79]C
#> 143 1056.40973 y8_   y_   8 1 RGSHKATC ARGS[79]HKAT[79]C
#> 144  161.01647 y1_   y_   1 1        C ARGS[79]HKAT[79]C
#> 145  341.06415 y2_   y_   2 1       TC ARGS[79]HKAT[79]C
#> 146  268.14041 b3*   b*   3 1      ARG ARGS[79]HKAT[79]C
#> 147  434.17244 b4*   b*   4 1     ARGS ARGS[79]HKAT[79]C
#> 148  571.23135 b5*   b*   5 1    ARGSH ARGS[79]HKAT[79]C
#> 149  699.32631 b6*   b*   6 1   ARGSHK ARGS[79]HKAT[79]C
#> 150  770.36342 b7*   b*   7 1  ARGSHKA ARGS[79]HKAT[79]C
#> 151  950.41110 b8*   b*   8 1 ARGSHKAT ARGS[79]HKAT[79]C
#> 152  541.18023 y4*   y*   4 1     KATC ARGS[79]HKAT[79]C
#> 153  678.23914 y5*   y*   5 1    HKATC ARGS[79]HKAT[79]C
#> 154  844.27117 y6*   y*   6 1   SHKATC ARGS[79]HKAT[79]C
#> 155  901.29263 y7*   y*   7 1  GSHKATC ARGS[79]HKAT[79]C
#> 156 1057.39374 y8*   y*   8 1 RGSHKATC ARGS[79]HKAT[79]C