Create an adjacency matrix.

This function created a peptide-by-protein adjacency matrix from a character or an instance of class PSM.

The character is formatted as x <- c("ProtA", "ProtB", "ProtA;ProtB", ...), as commonly encoutered in proteomics data spreadsheets. It defines that the first peptide is mapped to protein "ProtA", the second one to protein "ProtB", the third one to "ProtA" and "ProtB", and so on. The resulting matrix contain length(x) rows an as many columns as there are unique protein idenifiers in x. The column are always named after the protein idenifiers. If the protein identifier vector is named and the names are unique, these are then used to name to matrix rows.

makeAdjacencyMatrix(
  x,
  split = ";",
  peptides = "sequence",
  proteins = "DatabaseAccess",
  sparse = FALSE
)

Arguments

x
Either an instance of class PSM or a character. See example below for details.

split
character(1) defining how to split the string of protein identifiers.

peptides
character(1) indicating the name of the variable that defines peptides in the PSM object. Default is sequence.

proteins
character(1) indicating the name of the variable that defines proteins in the PSM object. Default is DatanbaseAccess.

sparse
logical(1) defining is a sparse adjacency matrix should be generated. Default is FALSE.

Value

A peptide-by-protein adjacency matrix.

Author

Laurent Gatto

Examples

## Protein vector without names
prots <- c("ProtA", "ProtB", "ProtA;ProtB")
makeAdjacencyMatrix(prots)
#>      ProtA ProtB
#> [1,]     1     0
#> [2,]     0     1
#> [3,]     1     1

## Named protein vector
names(prots) <- c("pep1", "pep2", "pep3")
makeAdjacencyMatrix(prots)
#>      ProtA ProtB
#> pep1     1     0
#> pep2     0     1
#> pep3     1     1