Calculate ions produced by fragmentation with variable modifications

This method calculates a-, b-, c-, x-, y- and z-ions produced by fragmentation.

Available methods

The default method with signature sequence = "character" and object = "missing" calculates the theoretical fragments for a peptide sequence. It returns a data.frame with the columns mz, ion, type, pos, z, seq and peptide.

Additional method can be defined that will adapt their behaviour based on spectra defined in object. See for example the MSnbase package that implements a method for objects of class Spectrum2.

# S4 method for class 'character,missing'
calculateFragments(
  sequence,
  type = c("b", "y"),
  z = 1,
  fixed_modifications = c(C = 57.02146),
  variable_modifications = numeric(),
  max_mods = Inf,
  neutralLoss = defaultNeutralLoss(),
  verbose = TRUE,
  modifications = NULL
)

Arguments

sequence

character() providing a peptide sequence.

type

character vector of target ions; possible values: c("a", "b", "c", "x", "y", "z"). Default is type = c("b", "y").

z

numeric with a desired charge state; default is 1.

fixed_modifications

A named numeric vector of used fixed modifications. The name must correspond to the one-letter-code of the modified amino acid and the numeric value must represent the mass that should be added to the original amino accid mass, default: Carbamidomethyl modifications = c(C = 57.02146). Use Nterm or Cterm as names for modifications that should be added to the amino respectively carboxyl-terminus.

variable_modifications

A named numeric vector of variable modifications. Depending on the maximum number of modifications (max_mods), all possible combinations are returned.

max_mods

A numeric indicating the maximum number of variable modifications allowed on the sequence at once. Does not include fixed modifications. Default value is positive infinity.

neutralLoss

list, it has to have two named elments, namely water and ammonia that contain a character vector which type of neutral loss should be calculated. Currently neutral loss on the C terminal "Cterm", at the amino acids c("D", "E", "S", "T") for "water" (shown with an _) and c("K", "N", "Q", "R") for "ammonia" (shown with an *) are supported.

There is a helper function `defaultNeutralLoss()` that returns
the correct list. It has two arguments `disableWaterLoss` and
`disableAmmoniaLoss` to remove single neutral loss options. See
the example section for use cases.

verbose

logical(1). If TRUE (default) the used modifications are printed.

modifications

Named numeric(). Deprecated modifications parameter. Will override fixed_modifications but is set to NULL by default. Please refrain from using it, opt for fixed_modifications instead.

Value

A data.frame showing all the ions produced by fragmentation with all possible combinations of modifications. The used variable modifications are displayed in the peptide column through the use of amino acids followed by the modification within brackets. Fixed modifications are not displayed.

Author

Sebastian Gibb mail@sebastiangibb.de

Guillaume Deflandre guillaume.deflandre@uclouvain.be

Examples

## General use
calculateFragments(sequence = "ARGSHKATC",
                   type = c("b", "y"), z = 1,
                   fixed_modifications = c(C = 57),
                   variable_modifications = c(S = 79, Y = 79, T = 79),
                   max_mods = 2)
#> Fixed modifications used: C=57
#> Variable modifications used: S=79, Y=79, T=79
#>             mz ion type pos z      seq           peptide
#> 1     72.04439  b1    b   1 1        A         ARGSHKATC
#> 2    228.14550  b2    b   2 1       AR         ARGSHKATC
#> 3    285.16696  b3    b   3 1      ARG         ARGSHKATC
#> 4    372.19899  b4    b   4 1     ARGS         ARGSHKATC
#> 5    509.25790  b5    b   5 1    ARGSH         ARGSHKATC
#> 6    637.35286  b6    b   6 1   ARGSHK         ARGSHKATC
#> 7    708.38997  b7    b   7 1  ARGSHKA         ARGSHKATC
#> 8    809.43765  b8    b   8 1 ARGSHKAT         ARGSHKATC
#> 9    179.02703  y1    y   1 1        C         ARGSHKATC
#> 10   280.07471  y2    y   2 1       TC         ARGSHKATC
#> 11   351.11182  y3    y   3 1      ATC         ARGSHKATC
#> 12   479.20678  y4    y   4 1     KATC         ARGSHKATC
#> 13   616.26569  y5    y   5 1    HKATC         ARGSHKATC
#> 14   703.29772  y6    y   6 1   SHKATC         ARGSHKATC
#> 15   760.31918  y7    y   7 1  GSHKATC         ARGSHKATC
#> 16   916.42029  y8    y   8 1 RGSHKATC         ARGSHKATC
#> 17   491.24733 b5_   b_   5 1    ARGSH         ARGSHKATC
#> 18   619.34229 b6_   b_   6 1   ARGSHK         ARGSHKATC
#> 19   690.37940 b7_   b_   7 1  ARGSHKA         ARGSHKATC
#> 20   791.42708 b8_   b_   8 1 ARGSHKAT         ARGSHKATC
#> 21   333.10126 y3_   y_   3 1      ATC         ARGSHKATC
#> 22   461.19622 y4_   y_   4 1     KATC         ARGSHKATC
#> 23   598.25513 y5_   y_   5 1    HKATC         ARGSHKATC
#> 24   685.28716 y6_   y_   6 1   SHKATC         ARGSHKATC
#> 25   742.30862 y7_   y_   7 1  GSHKATC         ARGSHKATC
#> 26   898.40973 y8_   y_   8 1 RGSHKATC         ARGSHKATC
#> 27   161.01647 y1_   y_   1 1        C         ARGSHKATC
#> 28   262.06415 y2_   y_   2 1       TC         ARGSHKATC
#> 29   268.14041 b3*   b*   3 1      ARG         ARGSHKATC
#> 30   355.17244 b4*   b*   4 1     ARGS         ARGSHKATC
#> 31   492.23135 b5*   b*   5 1    ARGSH         ARGSHKATC
#> 32   620.32631 b6*   b*   6 1   ARGSHK         ARGSHKATC
#> 33   691.36342 b7*   b*   7 1  ARGSHKA         ARGSHKATC
#> 34   792.41110 b8*   b*   8 1 ARGSHKAT         ARGSHKATC
#> 35   462.18023 y4*   y*   4 1     KATC         ARGSHKATC
#> 36   599.23914 y5*   y*   5 1    HKATC         ARGSHKATC
#> 37   686.27117 y6*   y*   6 1   SHKATC         ARGSHKATC
#> 38   743.29263 y7*   y*   7 1  GSHKATC         ARGSHKATC
#> 39   899.39374 y8*   y*   8 1 RGSHKATC         ARGSHKATC
#> 40    72.04439  b1    b   1 1        A     ARGS[79]HKATC
#> 41   228.14550  b2    b   2 1       AR     ARGS[79]HKATC
#> 42   285.16696  b3    b   3 1      ARG     ARGS[79]HKATC
#> 43   451.19899  b4    b   4 1     ARGS     ARGS[79]HKATC
#> 44   588.25790  b5    b   5 1    ARGSH     ARGS[79]HKATC
#> 45   716.35286  b6    b   6 1   ARGSHK     ARGS[79]HKATC
#> 46   787.38997  b7    b   7 1  ARGSHKA     ARGS[79]HKATC
#> 47   888.43765  b8    b   8 1 ARGSHKAT     ARGS[79]HKATC
#> 48   179.02703  y1    y   1 1        C     ARGS[79]HKATC
#> 49   280.07471  y2    y   2 1       TC     ARGS[79]HKATC
#> 50   351.11182  y3    y   3 1      ATC     ARGS[79]HKATC
#> 51   479.20678  y4    y   4 1     KATC     ARGS[79]HKATC
#> 52   616.26569  y5    y   5 1    HKATC     ARGS[79]HKATC
#> 53   782.29772  y6    y   6 1   SHKATC     ARGS[79]HKATC
#> 54   839.31918  y7    y   7 1  GSHKATC     ARGS[79]HKATC
#> 55   995.42029  y8    y   8 1 RGSHKATC     ARGS[79]HKATC
#> 56   570.24733 b5_   b_   5 1    ARGSH     ARGS[79]HKATC
#> 57   698.34229 b6_   b_   6 1   ARGSHK     ARGS[79]HKATC
#> 58   769.37940 b7_   b_   7 1  ARGSHKA     ARGS[79]HKATC
#> 59   870.42708 b8_   b_   8 1 ARGSHKAT     ARGS[79]HKATC
#> 60   333.10126 y3_   y_   3 1      ATC     ARGS[79]HKATC
#> 61   461.19622 y4_   y_   4 1     KATC     ARGS[79]HKATC
#> 62   598.25513 y5_   y_   5 1    HKATC     ARGS[79]HKATC
#> 63   764.28716 y6_   y_   6 1   SHKATC     ARGS[79]HKATC
#> 64   821.30862 y7_   y_   7 1  GSHKATC     ARGS[79]HKATC
#> 65   977.40973 y8_   y_   8 1 RGSHKATC     ARGS[79]HKATC
#> 66   161.01647 y1_   y_   1 1        C     ARGS[79]HKATC
#> 67   262.06415 y2_   y_   2 1       TC     ARGS[79]HKATC
#> 68   268.14041 b3*   b*   3 1      ARG     ARGS[79]HKATC
#> 69   434.17244 b4*   b*   4 1     ARGS     ARGS[79]HKATC
#> 70   571.23135 b5*   b*   5 1    ARGSH     ARGS[79]HKATC
#> 71   699.32631 b6*   b*   6 1   ARGSHK     ARGS[79]HKATC
#> 72   770.36342 b7*   b*   7 1  ARGSHKA     ARGS[79]HKATC
#> 73   871.41110 b8*   b*   8 1 ARGSHKAT     ARGS[79]HKATC
#> 74   462.18023 y4*   y*   4 1     KATC     ARGS[79]HKATC
#> 75   599.23914 y5*   y*   5 1    HKATC     ARGS[79]HKATC
#> 76   765.27117 y6*   y*   6 1   SHKATC     ARGS[79]HKATC
#> 77   822.29263 y7*   y*   7 1  GSHKATC     ARGS[79]HKATC
#> 78   978.39374 y8*   y*   8 1 RGSHKATC     ARGS[79]HKATC
#> 79    72.04439  b1    b   1 1        A     ARGSHKAT[79]C
#> 80   228.14550  b2    b   2 1       AR     ARGSHKAT[79]C
#> 81   285.16696  b3    b   3 1      ARG     ARGSHKAT[79]C
#> 82   372.19899  b4    b   4 1     ARGS     ARGSHKAT[79]C
#> 83   509.25790  b5    b   5 1    ARGSH     ARGSHKAT[79]C
#> 84   637.35286  b6    b   6 1   ARGSHK     ARGSHKAT[79]C
#> 85   708.38997  b7    b   7 1  ARGSHKA     ARGSHKAT[79]C
#> 86   888.43765  b8    b   8 1 ARGSHKAT     ARGSHKAT[79]C
#> 87   179.02703  y1    y   1 1        C     ARGSHKAT[79]C
#> 88   359.07471  y2    y   2 1       TC     ARGSHKAT[79]C
#> 89   430.11182  y3    y   3 1      ATC     ARGSHKAT[79]C
#> 90   558.20678  y4    y   4 1     KATC     ARGSHKAT[79]C
#> 91   695.26569  y5    y   5 1    HKATC     ARGSHKAT[79]C
#> 92   782.29772  y6    y   6 1   SHKATC     ARGSHKAT[79]C
#> 93   839.31918  y7    y   7 1  GSHKATC     ARGSHKAT[79]C
#> 94   995.42029  y8    y   8 1 RGSHKATC     ARGSHKAT[79]C
#> 95   491.24733 b5_   b_   5 1    ARGSH     ARGSHKAT[79]C
#> 96   619.34229 b6_   b_   6 1   ARGSHK     ARGSHKAT[79]C
#> 97   690.37940 b7_   b_   7 1  ARGSHKA     ARGSHKAT[79]C
#> 98   870.42708 b8_   b_   8 1 ARGSHKAT     ARGSHKAT[79]C
#> 99   412.10126 y3_   y_   3 1      ATC     ARGSHKAT[79]C
#> 100  540.19622 y4_   y_   4 1     KATC     ARGSHKAT[79]C
#> 101  677.25513 y5_   y_   5 1    HKATC     ARGSHKAT[79]C
#> 102  764.28716 y6_   y_   6 1   SHKATC     ARGSHKAT[79]C
#> 103  821.30862 y7_   y_   7 1  GSHKATC     ARGSHKAT[79]C
#> 104  977.40973 y8_   y_   8 1 RGSHKATC     ARGSHKAT[79]C
#> 105  161.01647 y1_   y_   1 1        C     ARGSHKAT[79]C
#> 106  341.06415 y2_   y_   2 1       TC     ARGSHKAT[79]C
#> 107  268.14041 b3*   b*   3 1      ARG     ARGSHKAT[79]C
#> 108  355.17244 b4*   b*   4 1     ARGS     ARGSHKAT[79]C
#> 109  492.23135 b5*   b*   5 1    ARGSH     ARGSHKAT[79]C
#> 110  620.32631 b6*   b*   6 1   ARGSHK     ARGSHKAT[79]C
#> 111  691.36342 b7*   b*   7 1  ARGSHKA     ARGSHKAT[79]C
#> 112  871.41110 b8*   b*   8 1 ARGSHKAT     ARGSHKAT[79]C
#> 113  541.18023 y4*   y*   4 1     KATC     ARGSHKAT[79]C
#> 114  678.23914 y5*   y*   5 1    HKATC     ARGSHKAT[79]C
#> 115  765.27117 y6*   y*   6 1   SHKATC     ARGSHKAT[79]C
#> 116  822.29263 y7*   y*   7 1  GSHKATC     ARGSHKAT[79]C
#> 117  978.39374 y8*   y*   8 1 RGSHKATC     ARGSHKAT[79]C
#> 118   72.04439  b1    b   1 1        A ARGS[79]HKAT[79]C
#> 119  228.14550  b2    b   2 1       AR ARGS[79]HKAT[79]C
#> 120  285.16696  b3    b   3 1      ARG ARGS[79]HKAT[79]C
#> 121  451.19899  b4    b   4 1     ARGS ARGS[79]HKAT[79]C
#> 122  588.25790  b5    b   5 1    ARGSH ARGS[79]HKAT[79]C
#> 123  716.35286  b6    b   6 1   ARGSHK ARGS[79]HKAT[79]C
#> 124  787.38997  b7    b   7 1  ARGSHKA ARGS[79]HKAT[79]C
#> 125  967.43765  b8    b   8 1 ARGSHKAT ARGS[79]HKAT[79]C
#> 126  179.02703  y1    y   1 1        C ARGS[79]HKAT[79]C
#> 127  359.07471  y2    y   2 1       TC ARGS[79]HKAT[79]C
#> 128  430.11182  y3    y   3 1      ATC ARGS[79]HKAT[79]C
#> 129  558.20678  y4    y   4 1     KATC ARGS[79]HKAT[79]C
#> 130  695.26569  y5    y   5 1    HKATC ARGS[79]HKAT[79]C
#> 131  861.29772  y6    y   6 1   SHKATC ARGS[79]HKAT[79]C
#> 132  918.31918  y7    y   7 1  GSHKATC ARGS[79]HKAT[79]C
#> 133 1074.42029  y8    y   8 1 RGSHKATC ARGS[79]HKAT[79]C
#> 134  570.24733 b5_   b_   5 1    ARGSH ARGS[79]HKAT[79]C
#> 135  698.34229 b6_   b_   6 1   ARGSHK ARGS[79]HKAT[79]C
#> 136  769.37940 b7_   b_   7 1  ARGSHKA ARGS[79]HKAT[79]C
#> 137  949.42708 b8_   b_   8 1 ARGSHKAT ARGS[79]HKAT[79]C
#> 138  412.10126 y3_   y_   3 1      ATC ARGS[79]HKAT[79]C
#> 139  540.19622 y4_   y_   4 1     KATC ARGS[79]HKAT[79]C
#> 140  677.25513 y5_   y_   5 1    HKATC ARGS[79]HKAT[79]C
#> 141  843.28716 y6_   y_   6 1   SHKATC ARGS[79]HKAT[79]C
#> 142  900.30862 y7_   y_   7 1  GSHKATC ARGS[79]HKAT[79]C
#> 143 1056.40973 y8_   y_   8 1 RGSHKATC ARGS[79]HKAT[79]C
#> 144  161.01647 y1_   y_   1 1        C ARGS[79]HKAT[79]C
#> 145  341.06415 y2_   y_   2 1       TC ARGS[79]HKAT[79]C
#> 146  268.14041 b3*   b*   3 1      ARG ARGS[79]HKAT[79]C
#> 147  434.17244 b4*   b*   4 1     ARGS ARGS[79]HKAT[79]C
#> 148  571.23135 b5*   b*   5 1    ARGSH ARGS[79]HKAT[79]C
#> 149  699.32631 b6*   b*   6 1   ARGSHK ARGS[79]HKAT[79]C
#> 150  770.36342 b7*   b*   7 1  ARGSHKA ARGS[79]HKAT[79]C
#> 151  950.41110 b8*   b*   8 1 ARGSHKAT ARGS[79]HKAT[79]C
#> 152  541.18023 y4*   y*   4 1     KATC ARGS[79]HKAT[79]C
#> 153  678.23914 y5*   y*   5 1    HKATC ARGS[79]HKAT[79]C
#> 154  844.27117 y6*   y*   6 1   SHKATC ARGS[79]HKAT[79]C
#> 155  901.29263 y7*   y*   7 1  GSHKATC ARGS[79]HKAT[79]C
#> 156 1057.39374 y8*   y*   8 1 RGSHKATC ARGS[79]HKAT[79]C

## calculate fragments for ACE with default modification
calculateFragments("ACE", fixed_modifications = c(C = 57.02146))
#> Fixed modifications used: C=57.02146
#> Variable modifications used: None
#>          mz ion type pos z seq peptide
#> 1  72.04439  b1    b   1 1   A     ACE
#> 2 232.07504  b2    b   2 1  AC     ACE
#> 3 148.06043  y1    y   1 1   E     ACE
#> 4 308.09108  y2    y   2 1  CE     ACE
#> 5 130.04987 y1_   y_   1 1   E     ACE
#> 6 290.08052 y2_   y_   2 1  CE     ACE

#' ## calculate fragments for ACE with an added variable modification
calculateFragments("ACE", variable_modifications = c(A = 43.25))
#> Fixed modifications used: C=57.02146
#> Variable modifications used: A=43.25
#>           mz ion type pos z seq    peptide
#> 1   72.04439  b1    b   1 1   A        ACE
#> 2  232.07504  b2    b   2 1  AC        ACE
#> 3  148.06043  y1    y   1 1   E        ACE
#> 4  308.09108  y2    y   2 1  CE        ACE
#> 5  130.04987 y1_   y_   1 1   E        ACE
#> 6  290.08052 y2_   y_   2 1  CE        ACE
#> 7  115.29439  b1    b   1 1   A A[43.25]CE
#> 8  275.32504  b2    b   2 1  AC A[43.25]CE
#> 9  148.06043  y1    y   1 1   E A[43.25]CE
#> 10 308.09108  y2    y   2 1  CE A[43.25]CE
#> 11 130.04987 y1_   y_   1 1   E A[43.25]CE
#> 12 290.08052 y2_   y_   2 1  CE A[43.25]CE

## calculate fragments for ACE with an added N-terminal modification
calculateFragments("ACE", fixed_modifications = c(C = 57.02146, Nterm = 229.1629))
#> Fixed modifications used: C=57.02146, Nterm=229.1629
#> Variable modifications used: None
#>         mz ion type pos z seq peptide
#> 1 301.2073  b1    b   1 1   A     ACE
#> 2 461.2379  b2    b   2 1  AC     ACE
#> 3 148.0604  y1    y   1 1   E     ACE
#> 4 308.0911  y2    y   2 1  CE     ACE
#> 5 130.0499 y1_   y_   1 1   E     ACE
#> 6 290.0805 y2_   y_   2 1  CE     ACE

## calculate fragments for ACE without any modifications
calculateFragments("ACE", fixed_modifications = NULL)
#> Fixed modifications used: None
#> Variable modifications used: None
#>          mz ion type pos z seq peptide
#> 1  72.04439  b1    b   1 1   A     ACE
#> 2 175.05358  b2    b   2 1  AC     ACE
#> 3 148.06043  y1    y   1 1   E     ACE
#> 4 251.06962  y2    y   2 1  CE     ACE
#> 5 130.04987 y1_   y_   1 1   E     ACE
#> 6 233.05906 y2_   y_   2 1  CE     ACE

calculateFragments("VESITARHGEVLQLRPK",
                   type = c("a", "b", "c", "x", "y", "z"),
                   z = 1:2)
#> Fixed modifications used: C=57.02146
#> Variable modifications used: None
#>             mz  ion type pos z              seq           peptide
#> 1     72.08077   a1    a   1 1                V VESITARHGEVLQLRPK
#> 2     36.54402   a1    a   1 2                V VESITARHGEVLQLRPK
#> 3    201.12336   a2    a   2 1               VE VESITARHGEVLQLRPK
#> 4    101.06532   a2    a   2 2               VE VESITARHGEVLQLRPK
#> 5    288.15539   a3    a   3 1              VES VESITARHGEVLQLRPK
#> 6    144.58133   a3    a   3 2              VES VESITARHGEVLQLRPK
#> 7    401.23945   a4    a   4 1             VESI VESITARHGEVLQLRPK
#> 8    201.12336   a4    a   4 2             VESI VESITARHGEVLQLRPK
#> 9    502.28713   a5    a   5 1            VESIT VESITARHGEVLQLRPK
#> 10   251.64720   a5    a   5 2            VESIT VESITARHGEVLQLRPK
#> 11   573.32424   a6    a   6 1           VESITA VESITARHGEVLQLRPK
#> 12   287.16576   a6    a   6 2           VESITA VESITARHGEVLQLRPK
#> 13   729.42535   a7    a   7 1          VESITAR VESITARHGEVLQLRPK
#> 14   365.21631   a7    a   7 2          VESITAR VESITARHGEVLQLRPK
#> 15   866.48426   a8    a   8 1         VESITARH VESITARHGEVLQLRPK
#> 16   433.74577   a8    a   8 2         VESITARH VESITARHGEVLQLRPK
#> 17   923.50572   a9    a   9 1        VESITARHG VESITARHGEVLQLRPK
#> 18   462.25650   a9    a   9 2        VESITARHG VESITARHGEVLQLRPK
#> 19  1052.54831  a10    a  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 20   526.77779  a10    a  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 21  1151.61672  a11    a  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 22   576.31200  a11    a  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 23  1264.70078  a12    a  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 24   632.85403  a12    a  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 25  1392.75936  a13    a  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 26   696.88332  a13    a  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 27  1505.84342  a14    a  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 28   753.42535  a14    a  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 29  1661.94453  a15    a  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 30   831.47590  a15    a  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 31  1758.99729  a16    a  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 32   880.00228  a16    a  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 33   100.07569   b1    b   1 1                V VESITARHGEVLQLRPK
#> 34    50.54148   b1    b   1 2                V VESITARHGEVLQLRPK
#> 35   229.11828   b2    b   2 1               VE VESITARHGEVLQLRPK
#> 36   115.06278   b2    b   2 2               VE VESITARHGEVLQLRPK
#> 37   316.15031   b3    b   3 1              VES VESITARHGEVLQLRPK
#> 38   158.57879   b3    b   3 2              VES VESITARHGEVLQLRPK
#> 39   429.23437   b4    b   4 1             VESI VESITARHGEVLQLRPK
#> 40   215.12082   b4    b   4 2             VESI VESITARHGEVLQLRPK
#> 41   530.28205   b5    b   5 1            VESIT VESITARHGEVLQLRPK
#> 42   265.64466   b5    b   5 2            VESIT VESITARHGEVLQLRPK
#> 43   601.31916   b6    b   6 1           VESITA VESITARHGEVLQLRPK
#> 44   301.16322   b6    b   6 2           VESITA VESITARHGEVLQLRPK
#> 45   757.42027   b7    b   7 1          VESITAR VESITARHGEVLQLRPK
#> 46   379.21377   b7    b   7 2          VESITAR VESITARHGEVLQLRPK
#> 47   894.47918   b8    b   8 1         VESITARH VESITARHGEVLQLRPK
#> 48   447.74323   b8    b   8 2         VESITARH VESITARHGEVLQLRPK
#> 49   951.50064   b9    b   9 1        VESITARHG VESITARHGEVLQLRPK
#> 50   476.25396   b9    b   9 2        VESITARHG VESITARHGEVLQLRPK
#> 51  1080.54323  b10    b  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 52   540.77525  b10    b  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 53  1179.61164  b11    b  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 54   590.30946  b11    b  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 55  1292.69570  b12    b  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 56   646.85149  b12    b  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 57  1420.75428  b13    b  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 58   710.88078  b13    b  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 59  1533.83834  b14    b  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 60   767.42281  b14    b  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 61  1689.93945  b15    b  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 62   845.47336  b15    b  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 63  1786.99221  b16    b  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 64   893.99974  b16    b  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 65   117.10224   c1    c   1 1                V VESITARHGEVLQLRPK
#> 66    59.05476   c1    c   1 2                V VESITARHGEVLQLRPK
#> 67   246.14482   c2    c   2 1               VE VESITARHGEVLQLRPK
#> 68   123.57605   c2    c   2 2               VE VESITARHGEVLQLRPK
#> 69   333.17685   c3    c   3 1              VES VESITARHGEVLQLRPK
#> 70   167.09207   c3    c   3 2              VES VESITARHGEVLQLRPK
#> 71   446.26091   c4    c   4 1             VESI VESITARHGEVLQLRPK
#> 72   223.63410   c4    c   4 2             VESI VESITARHGEVLQLRPK
#> 73   547.30860   c5    c   5 1            VESIT VESITARHGEVLQLRPK
#> 74   274.15794   c5    c   5 2            VESIT VESITARHGEVLQLRPK
#> 75   618.34571   c6    c   6 1           VESITA VESITARHGEVLQLRPK
#> 76   309.67649   c6    c   6 2           VESITA VESITARHGEVLQLRPK
#> 77   774.44682   c7    c   7 1          VESITAR VESITARHGEVLQLRPK
#> 78   387.72705   c7    c   7 2          VESITAR VESITARHGEVLQLRPK
#> 79   911.50573   c8    c   8 1         VESITARH VESITARHGEVLQLRPK
#> 80   456.25650   c8    c   8 2         VESITARH VESITARHGEVLQLRPK
#> 81   968.52719   c9    c   9 1        VESITARHG VESITARHGEVLQLRPK
#> 82   484.76723   c9    c   9 2        VESITARHG VESITARHGEVLQLRPK
#> 83  1097.56977  c10    c  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 84   549.28853  c10    c  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 85  1196.63819  c11    c  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 86   598.82273  c11    c  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 87  1309.72224  c12    c  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 88   655.36476  c12    c  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 89  1437.78083  c13    c  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 90   719.39405  c13    c  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 91  1550.86488  c14    c  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 92   775.93608  c14    c  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 93  1706.96600  c15    c  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 94   853.98664  c15    c  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 95  1804.01876  c16    c  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 96   902.51302  c16    c  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 97   173.09207   x1    x   1 1                K VESITARHGEVLQLRPK
#> 98    87.04967   x1    x   1 2                K VESITARHGEVLQLRPK
#> 99   270.14483   x2    x   2 1               PK VESITARHGEVLQLRPK
#> 100  135.57605   x2    x   2 2               PK VESITARHGEVLQLRPK
#> 101  426.24594   x3    x   3 1              RPK VESITARHGEVLQLRPK
#> 102  213.62661   x3    x   3 2              RPK VESITARHGEVLQLRPK
#> 103  539.33000   x4    x   4 1             LRPK VESITARHGEVLQLRPK
#> 104  270.16864   x4    x   4 2             LRPK VESITARHGEVLQLRPK
#> 105  667.38858   x5    x   5 1            QLRPK VESITARHGEVLQLRPK
#> 106  334.19793   x5    x   5 2            QLRPK VESITARHGEVLQLRPK
#> 107  780.47264   x6    x   6 1           LQLRPK VESITARHGEVLQLRPK
#> 108  390.73996   x6    x   6 2           LQLRPK VESITARHGEVLQLRPK
#> 109  879.54105   x7    x   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 110  440.27416   x7    x   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 111 1008.58364   x8    x   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 112  504.79546   x8    x   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 113 1065.60510   x9    x   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 114  533.30619   x9    x   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 115 1202.66401  x10    x  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 116  601.83564  x10    x  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 117 1358.76512  x11    x  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 118  679.88620  x11    x  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 119 1429.80223  x12    x  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 120  715.40475  x12    x  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 121 1530.84991  x13    x  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 122  765.92859  x13    x  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 123 1643.93397  x14    x  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 124  822.47062  x14    x  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 125 1730.96600  x15    x  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 126  865.98664  x15    x  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 127 1860.00859  x16    x  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 128  930.50793  x16    x  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 129  147.11280   y1    y   1 1                K VESITARHGEVLQLRPK
#> 130   74.06004   y1    y   1 2                K VESITARHGEVLQLRPK
#> 131  244.16556   y2    y   2 1               PK VESITARHGEVLQLRPK
#> 132  122.58642   y2    y   2 2               PK VESITARHGEVLQLRPK
#> 133  400.26667   y3    y   3 1              RPK VESITARHGEVLQLRPK
#> 134  200.63697   y3    y   3 2              RPK VESITARHGEVLQLRPK
#> 135  513.35073   y4    y   4 1             LRPK VESITARHGEVLQLRPK
#> 136  257.17900   y4    y   4 2             LRPK VESITARHGEVLQLRPK
#> 137  641.40931   y5    y   5 1            QLRPK VESITARHGEVLQLRPK
#> 138  321.20829   y5    y   5 2            QLRPK VESITARHGEVLQLRPK
#> 139  754.49337   y6    y   6 1           LQLRPK VESITARHGEVLQLRPK
#> 140  377.75032   y6    y   6 2           LQLRPK VESITARHGEVLQLRPK
#> 141  853.56178   y7    y   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 142  427.28453   y7    y   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 143  982.60437   y8    y   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 144  491.80582   y8    y   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 145 1039.62583   y9    y   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 146  520.31655   y9    y   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 147 1176.68474  y10    y  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 148  588.84601  y10    y  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 149 1332.78585  y11    y  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 150  666.89656  y11    y  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 151 1403.82296  y12    y  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 152  702.41512  y12    y  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 153 1504.87064  y13    y  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 154  752.93896  y13    y  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 155 1617.95470  y14    y  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 156  809.48099  y14    y  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 157 1704.98673  y15    y  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 158  852.99700  y15    y  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 159 1834.02932  y16    y  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 160  917.51830  y16    y  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 161  130.08625   z1    z   1 1                K VESITARHGEVLQLRPK
#> 162   65.54676   z1    z   1 2                K VESITARHGEVLQLRPK
#> 163  227.13901   z2    z   2 1               PK VESITARHGEVLQLRPK
#> 164  114.07314   z2    z   2 2               PK VESITARHGEVLQLRPK
#> 165  383.24012   z3    z   3 1              RPK VESITARHGEVLQLRPK
#> 166  192.12370   z3    z   3 2              RPK VESITARHGEVLQLRPK
#> 167  496.32418   z4    z   4 1             LRPK VESITARHGEVLQLRPK
#> 168  248.66573   z4    z   4 2             LRPK VESITARHGEVLQLRPK
#> 169  624.38276   z5    z   5 1            QLRPK VESITARHGEVLQLRPK
#> 170  312.69502   z5    z   5 2            QLRPK VESITARHGEVLQLRPK
#> 171  737.46682   z6    z   6 1           LQLRPK VESITARHGEVLQLRPK
#> 172  369.23705   z6    z   6 2           LQLRPK VESITARHGEVLQLRPK
#> 173  836.53523   z7    z   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 174  418.77125   z7    z   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 175  965.57782   z8    z   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 176  483.29255   z8    z   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 177 1022.59928   z9    z   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 178  511.80328   z9    z   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 179 1159.65819  z10    z  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 180  580.33273  z10    z  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 181 1315.75930  z11    z  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 182  658.38329  z11    z  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 183 1386.79641  z12    z  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 184  693.90184  z12    z  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 185 1487.84409  z13    z  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 186  744.42568  z13    z  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 187 1600.92815  z14    z  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 188  800.96771  z14    z  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 189 1687.96018  z15    z  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 190  844.48373  z15    z  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 191 1817.00277  z16    z  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 192  909.00502  z16    z  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 193  383.22889  a4_   a_   4 1             VESI VESITARHGEVLQLRPK
#> 194  192.11808  a4_   a_   4 2             VESI VESITARHGEVLQLRPK
#> 195  484.27657  a5_   a_   5 1            VESIT VESITARHGEVLQLRPK
#> 196  242.64192  a5_   a_   5 2            VESIT VESITARHGEVLQLRPK
#> 197  555.31368  a6_   a_   6 1           VESITA VESITARHGEVLQLRPK
#> 198  278.16048  a6_   a_   6 2           VESITA VESITARHGEVLQLRPK
#> 199  711.41479  a7_   a_   7 1          VESITAR VESITARHGEVLQLRPK
#> 200  356.21103  a7_   a_   7 2          VESITAR VESITARHGEVLQLRPK
#> 201  848.47370  a8_   a_   8 1         VESITARH VESITARHGEVLQLRPK
#> 202  424.74049  a8_   a_   8 2         VESITARH VESITARHGEVLQLRPK
#> 203  905.49516  a9_   a_   9 1        VESITARHG VESITARHGEVLQLRPK
#> 204  453.25122  a9_   a_   9 2        VESITARHG VESITARHGEVLQLRPK
#> 205 1034.53775 a10_   a_  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 206  517.77251 a10_   a_  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 207 1133.60616 a11_   a_  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 208  567.30672 a11_   a_  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 209 1246.69022 a12_   a_  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 210  623.84875 a12_   a_  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 211 1374.74880 a13_   a_  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 212  687.87804 a13_   a_  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 213 1487.83286 a14_   a_  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 214  744.42007 a14_   a_  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 215 1643.93397 a15_   a_  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 216  822.47062 a15_   a_  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 217 1740.98673 a16_   a_  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 218  870.99700 a16_   a_  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 219  411.22380  b4_   b_   4 1             VESI VESITARHGEVLQLRPK
#> 220  206.11554  b4_   b_   4 2             VESI VESITARHGEVLQLRPK
#> 221  512.27148  b5_   b_   5 1            VESIT VESITARHGEVLQLRPK
#> 222  256.63938  b5_   b_   5 2            VESIT VESITARHGEVLQLRPK
#> 223  583.30859  b6_   b_   6 1           VESITA VESITARHGEVLQLRPK
#> 224  292.15793  b6_   b_   6 2           VESITA VESITARHGEVLQLRPK
#> 225  739.40970  b7_   b_   7 1          VESITAR VESITARHGEVLQLRPK
#> 226  370.20849  b7_   b_   7 2          VESITAR VESITARHGEVLQLRPK
#> 227  876.46861  b8_   b_   8 1         VESITARH VESITARHGEVLQLRPK
#> 228  438.73794  b8_   b_   8 2         VESITARH VESITARHGEVLQLRPK
#> 229  933.49007  b9_   b_   9 1        VESITARHG VESITARHGEVLQLRPK
#> 230  467.24867  b9_   b_   9 2        VESITARHG VESITARHGEVLQLRPK
#> 231 1062.53266 b10_   b_  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 232  531.76997 b10_   b_  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 233 1161.60107 b11_   b_  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 234  581.30417 b11_   b_  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 235 1274.68513 b12_   b_  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 236  637.84620 b12_   b_  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 237 1402.74371 b13_   b_  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 238  701.87549 b13_   b_  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 239 1515.82777 b14_   b_  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 240  758.41752 b14_   b_  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 241 1671.92888 b15_   b_  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 242  836.46808 b15_   b_  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 243 1768.98164 b16_   b_  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 244  884.99446 b16_   b_  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 245  428.25035  c4_   c_   4 1             VESI VESITARHGEVLQLRPK
#> 246  214.62881  c4_   c_   4 2             VESI VESITARHGEVLQLRPK
#> 247  529.29803  c5_   c_   5 1            VESIT VESITARHGEVLQLRPK
#> 248  265.15265  c5_   c_   5 2            VESIT VESITARHGEVLQLRPK
#> 249  600.33514  c6_   c_   6 1           VESITA VESITARHGEVLQLRPK
#> 250  300.67121  c6_   c_   6 2           VESITA VESITARHGEVLQLRPK
#> 251  756.43625  c7_   c_   7 1          VESITAR VESITARHGEVLQLRPK
#> 252  378.72176  c7_   c_   7 2          VESITAR VESITARHGEVLQLRPK
#> 253  893.49516  c8_   c_   8 1         VESITARH VESITARHGEVLQLRPK
#> 254  447.25122  c8_   c_   8 2         VESITARH VESITARHGEVLQLRPK
#> 255  950.51662  c9_   c_   9 1        VESITARHG VESITARHGEVLQLRPK
#> 256  475.76195  c9_   c_   9 2        VESITARHG VESITARHGEVLQLRPK
#> 257 1079.55921 c10_   c_  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 258  540.28324 c10_   c_  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 259 1178.62762 c11_   c_  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 260  589.81745 c11_   c_  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 261 1291.71168 c12_   c_  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 262  646.35948 c12_   c_  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 263 1419.77026 c13_   c_  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 264  710.38877 c13_   c_  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 265 1532.85432 c14_   c_  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 266  766.93080 c14_   c_  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 267 1688.95543 c15_   c_  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 268  844.98135 c15_   c_  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 269 1786.00819 c16_   c_  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 270  893.50773 c16_   c_  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 271  990.57307  x8_   x_   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 272  495.79017  x8_   x_   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 273 1625.92340 x14_   x_  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 274  813.46534 x14_   x_  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 275 1712.95543 x15_   x_  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 276  856.98135 x15_   x_  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 277 1841.99802 x16_   x_  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 278  921.50265 x16_   x_  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 279  964.59381  y8_   y_   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 280  482.80054  y8_   y_   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 281 1599.94414 y14_   y_  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 282  800.47571 y14_   y_  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 283 1686.97617 y15_   y_  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 284  843.99172 y15_   y_  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 285 1816.01876 y16_   y_  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 286  908.51302 y16_   y_  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 287  947.56726  z8_   z_   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 288  474.28727  z8_   z_   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 289 1582.91759 z14_   z_  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 290  791.96243 z14_   z_  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 291 1669.94962 z15_   z_  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 292  835.47845 z15_   z_  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 293 1798.99221 z16_   z_  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 294  899.99974 z16_   z_  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 295  155.08150  x1_   x_   1 1                K VESITARHGEVLQLRPK
#> 296   78.04439  x1_   x_   1 2                K VESITARHGEVLQLRPK
#> 297  252.13426  x2_   x_   2 1               PK VESITARHGEVLQLRPK
#> 298  126.57077  x2_   x_   2 2               PK VESITARHGEVLQLRPK
#> 299  408.23537  x3_   x_   3 1              RPK VESITARHGEVLQLRPK
#> 300  204.62132  x3_   x_   3 2              RPK VESITARHGEVLQLRPK
#> 301  521.31943  x4_   x_   4 1             LRPK VESITARHGEVLQLRPK
#> 302  261.16335  x4_   x_   4 2             LRPK VESITARHGEVLQLRPK
#> 303  649.37801  x5_   x_   5 1            QLRPK VESITARHGEVLQLRPK
#> 304  325.19264  x5_   x_   5 2            QLRPK VESITARHGEVLQLRPK
#> 305  762.46207  x6_   x_   6 1           LQLRPK VESITARHGEVLQLRPK
#> 306  381.73467  x6_   x_   6 2           LQLRPK VESITARHGEVLQLRPK
#> 307  861.53048  x7_   x_   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 308  431.26888  x7_   x_   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 309 1047.59453  x9_   x_   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 310  524.30090  x9_   x_   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 311 1184.65344 x10_   x_  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 312  592.83036 x10_   x_  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 313 1340.75455 x11_   x_  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 314  670.88091 x11_   x_  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 315 1411.79166 x12_   x_  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 316  706.39947 x12_   x_  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 317 1512.83934 x13_   x_  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 318  756.92331 x13_   x_  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 319  129.10224  y1_   y_   1 1                K VESITARHGEVLQLRPK
#> 320   65.05476  y1_   y_   1 2                K VESITARHGEVLQLRPK
#> 321  226.15500  y2_   y_   2 1               PK VESITARHGEVLQLRPK
#> 322  113.58114  y2_   y_   2 2               PK VESITARHGEVLQLRPK
#> 323  382.25611  y3_   y_   3 1              RPK VESITARHGEVLQLRPK
#> 324  191.63169  y3_   y_   3 2              RPK VESITARHGEVLQLRPK
#> 325  495.34017  y4_   y_   4 1             LRPK VESITARHGEVLQLRPK
#> 326  248.17372  y4_   y_   4 2             LRPK VESITARHGEVLQLRPK
#> 327  623.39875  y5_   y_   5 1            QLRPK VESITARHGEVLQLRPK
#> 328  312.20301  y5_   y_   5 2            QLRPK VESITARHGEVLQLRPK
#> 329  736.48281  y6_   y_   6 1           LQLRPK VESITARHGEVLQLRPK
#> 330  368.74504  y6_   y_   6 2           LQLRPK VESITARHGEVLQLRPK
#> 331  835.55122  y7_   y_   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 332  418.27925  y7_   y_   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 333 1021.61527  y9_   y_   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 334  511.31127  y9_   y_   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 335 1158.67418 y10_   y_  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 336  579.84073 y10_   y_  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 337 1314.77529 y11_   y_  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 338  657.89128 y11_   y_  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 339 1385.81240 y12_   y_  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 340  693.40984 y12_   y_  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 341 1486.86008 y13_   y_  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 342  743.93368 y13_   y_  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 343  112.07569  z1_   z_   1 1                K VESITARHGEVLQLRPK
#> 344   56.54148  z1_   z_   1 2                K VESITARHGEVLQLRPK
#> 345  209.12845  z2_   z_   2 1               PK VESITARHGEVLQLRPK
#> 346  105.06786  z2_   z_   2 2               PK VESITARHGEVLQLRPK
#> 347  365.22956  z3_   z_   3 1              RPK VESITARHGEVLQLRPK
#> 348  183.11842  z3_   z_   3 2              RPK VESITARHGEVLQLRPK
#> 349  478.31362  z4_   z_   4 1             LRPK VESITARHGEVLQLRPK
#> 350  239.66045  z4_   z_   4 2             LRPK VESITARHGEVLQLRPK
#> 351  606.37220  z5_   z_   5 1            QLRPK VESITARHGEVLQLRPK
#> 352  303.68974  z5_   z_   5 2            QLRPK VESITARHGEVLQLRPK
#> 353  719.45626  z6_   z_   6 1           LQLRPK VESITARHGEVLQLRPK
#> 354  360.23177  z6_   z_   6 2           LQLRPK VESITARHGEVLQLRPK
#> 355  818.52467  z7_   z_   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 356  409.76597  z7_   z_   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 357 1004.58872  z9_   z_   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 358  502.79800  z9_   z_   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 359 1141.64763 z10_   z_  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 360  571.32745 z10_   z_  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 361 1297.74874 z11_   z_  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 362  649.37801 z11_   z_  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 363 1368.78585 z12_   z_  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 364  684.89656 z12_   z_  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 365 1469.83353 z13_   z_  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 366  735.42040 z13_   z_  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 367  849.45771  a8*   a*   8 1         VESITARH VESITARHGEVLQLRPK
#> 368  425.23249  a8*   a*   8 2         VESITARH VESITARHGEVLQLRPK
#> 369  906.47917  a9*   a*   9 1        VESITARHG VESITARHGEVLQLRPK
#> 370  453.74322  a9*   a*   9 2        VESITARHG VESITARHGEVLQLRPK
#> 371 1035.52176 a10*   a*  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 372  518.26452 a10*   a*  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 373 1134.59017 a11*   a*  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 374  567.79872 a11*   a*  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 375 1247.67423 a12*   a*  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 376  624.34075 a12*   a*  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 377 1375.73281 a13*   a*  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 378  688.37004 a13*   a*  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 379 1488.81687 a14*   a*  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 380  744.91207 a14*   a*  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 381 1644.91798 a15*   a*  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 382  822.96263 a15*   a*  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 383 1741.97074 a16*   a*  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 384  871.48901 a16*   a*  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 385  877.45263  b8*   b*   8 1         VESITARH VESITARHGEVLQLRPK
#> 386  439.22995  b8*   b*   8 2         VESITARH VESITARHGEVLQLRPK
#> 387  934.47409  b9*   b*   9 1        VESITARHG VESITARHGEVLQLRPK
#> 388  467.74068  b9*   b*   9 2        VESITARHG VESITARHGEVLQLRPK
#> 389 1063.51668 b10*   b*  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 390  532.26198 b10*   b*  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 391 1162.58509 b11*   b*  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 392  581.79618 b11*   b*  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 393 1275.66915 b12*   b*  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 394  638.33821 b12*   b*  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 395 1403.72773 b13*   b*  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 396  702.36750 b13*   b*  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 397 1516.81179 b14*   b*  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 398  758.90953 b14*   b*  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 399 1672.91290 b15*   b*  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 400  836.96009 b15*   b*  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 401 1769.96566 b16*   b*  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 402  885.48647 b16*   b*  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 403  894.47918  c8*   c*   8 1         VESITARH VESITARHGEVLQLRPK
#> 404  447.74323  c8*   c*   8 2         VESITARH VESITARHGEVLQLRPK
#> 405  951.50064  c9*   c*   9 1        VESITARHG VESITARHGEVLQLRPK
#> 406  476.25396  c9*   c*   9 2        VESITARHG VESITARHGEVLQLRPK
#> 407 1080.54323 c10*   c*  10 1       VESITARHGE VESITARHGEVLQLRPK
#> 408  540.77525 c10*   c*  10 2       VESITARHGE VESITARHGEVLQLRPK
#> 409 1179.61164 c11*   c*  11 1      VESITARHGEV VESITARHGEVLQLRPK
#> 410  590.30946 c11*   c*  11 2      VESITARHGEV VESITARHGEVLQLRPK
#> 411 1292.69570 c12*   c*  12 1     VESITARHGEVL VESITARHGEVLQLRPK
#> 412  646.85149 c12*   c*  12 2     VESITARHGEVL VESITARHGEVLQLRPK
#> 413 1420.75428 c13*   c*  13 1    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 414  710.88078 c13*   c*  13 2    VESITARHGEVLQ VESITARHGEVLQLRPK
#> 415 1533.83834 c14*   c*  14 1   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 416  767.42281 c14*   c*  14 2   VESITARHGEVLQL VESITARHGEVLQLRPK
#> 417 1689.93945 c15*   c*  15 1  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 418  845.47336 c15*   c*  15 2  VESITARHGEVLQLR VESITARHGEVLQLRPK
#> 419 1786.99221 c16*   c*  16 1 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 420  893.99974 c16*   c*  16 2 VESITARHGEVLQLRP VESITARHGEVLQLRPK
#> 421  522.30345  x4*   x*   4 1             LRPK VESITARHGEVLQLRPK
#> 422  261.65536  x4*   x*   4 2             LRPK VESITARHGEVLQLRPK
#> 423  650.36203  x5*   x*   5 1            QLRPK VESITARHGEVLQLRPK
#> 424  325.68465  x5*   x*   5 2            QLRPK VESITARHGEVLQLRPK
#> 425  763.44609  x6*   x*   6 1           LQLRPK VESITARHGEVLQLRPK
#> 426  382.22668  x6*   x*   6 2           LQLRPK VESITARHGEVLQLRPK
#> 427  862.51450  x7*   x*   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 428  431.76089  x7*   x*   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 429  991.55709  x8*   x*   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 430  496.28218  x8*   x*   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 431 1048.57855  x9*   x*   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 432  524.79291  x9*   x*   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 433 1185.63746 x10*   x*  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 434  593.32237 x10*   x*  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 435 1341.73857 x11*   x*  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 436  671.37292 x11*   x*  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 437 1412.77568 x12*   x*  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 438  706.89148 x12*   x*  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 439 1513.82336 x13*   x*  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 440  757.41532 x13*   x*  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 441 1626.90742 x14*   x*  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 442  813.95735 x14*   x*  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 443 1713.93945 x15*   x*  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 444  857.47336 x15*   x*  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 445 1842.98204 x16*   x*  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 446  921.99466 x16*   x*  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 447  496.32418  y4*   y*   4 1             LRPK VESITARHGEVLQLRPK
#> 448  248.66573  y4*   y*   4 2             LRPK VESITARHGEVLQLRPK
#> 449  624.38276  y5*   y*   5 1            QLRPK VESITARHGEVLQLRPK
#> 450  312.69502  y5*   y*   5 2            QLRPK VESITARHGEVLQLRPK
#> 451  737.46682  y6*   y*   6 1           LQLRPK VESITARHGEVLQLRPK
#> 452  369.23705  y6*   y*   6 2           LQLRPK VESITARHGEVLQLRPK
#> 453  836.53523  y7*   y*   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 454  418.77125  y7*   y*   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 455  965.57782  y8*   y*   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 456  483.29255  y8*   y*   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 457 1022.59928  y9*   y*   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 458  511.80328  y9*   y*   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 459 1159.65819 y10*   y*  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 460  580.33273 y10*   y*  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 461 1315.75930 y11*   y*  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 462  658.38329 y11*   y*  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 463 1386.79641 y12*   y*  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 464  693.90184 y12*   y*  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 465 1487.84409 y13*   y*  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 466  744.42568 y13*   y*  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 467 1600.92815 y14*   y*  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 468  800.96771 y14*   y*  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 469 1687.96018 y15*   y*  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 470  844.48373 y15*   y*  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 471 1817.00277 y16*   y*  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 472  909.00502 y16*   y*  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 473  479.29763  z4*   z*   4 1             LRPK VESITARHGEVLQLRPK
#> 474  240.15245  z4*   z*   4 2             LRPK VESITARHGEVLQLRPK
#> 475  607.35621  z5*   z*   5 1            QLRPK VESITARHGEVLQLRPK
#> 476  304.18174  z5*   z*   5 2            QLRPK VESITARHGEVLQLRPK
#> 477  720.44027  z6*   z*   6 1           LQLRPK VESITARHGEVLQLRPK
#> 478  360.72377  z6*   z*   6 2           LQLRPK VESITARHGEVLQLRPK
#> 479  819.50868  z7*   z*   7 1          VLQLRPK VESITARHGEVLQLRPK
#> 480  410.25798  z7*   z*   7 2          VLQLRPK VESITARHGEVLQLRPK
#> 481  948.55127  z8*   z*   8 1         EVLQLRPK VESITARHGEVLQLRPK
#> 482  474.77927  z8*   z*   8 2         EVLQLRPK VESITARHGEVLQLRPK
#> 483 1005.57273  z9*   z*   9 1        GEVLQLRPK VESITARHGEVLQLRPK
#> 484  503.29000  z9*   z*   9 2        GEVLQLRPK VESITARHGEVLQLRPK
#> 485 1142.63164 z10*   z*  10 1       HGEVLQLRPK VESITARHGEVLQLRPK
#> 486  571.81946 z10*   z*  10 2       HGEVLQLRPK VESITARHGEVLQLRPK
#> 487 1298.73275 z11*   z*  11 1      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 488  649.87001 z11*   z*  11 2      RHGEVLQLRPK VESITARHGEVLQLRPK
#> 489 1369.76986 z12*   z*  12 1     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 490  685.38857 z12*   z*  12 2     ARHGEVLQLRPK VESITARHGEVLQLRPK
#> 491 1470.81754 z13*   z*  13 1    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 492  735.91241 z13*   z*  13 2    TARHGEVLQLRPK VESITARHGEVLQLRPK
#> 493 1583.90160 z14*   z*  14 1   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 494  792.45444 z14*   z*  14 2   ITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 495 1670.93363 z15*   z*  15 1  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 496  835.97045 z15*   z*  15 2  SITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 497 1799.97622 z16*   z*  16 1 ESITARHGEVLQLRPK VESITARHGEVLQLRPK
#> 498  900.49175 z16*   z*  16 2 ESITARHGEVLQLRPK VESITARHGEVLQLRPK

## neutral loss
defaultNeutralLoss()
#> $water
#> [1] "Cterm" "D"     "E"     "S"     "T"    
#> 
#> $ammonia
#> [1] "K" "N" "Q" "R"
#> 

## disable water loss on the C terminal
defaultNeutralLoss(disableWaterLoss="Cterm")
#> $water
#> [1] "D" "E" "S" "T"
#> 
#> $ammonia
#> [1] "K" "N" "Q" "R"
#> 

## real example
calculateFragments("PQR")
#> Fixed modifications used: C=57.02146
#> Variable modifications used: None
#>          mz ion type pos z seq peptide
#> 1  98.06004  b1    b   1 1   P     PQR
#> 2 226.11862  b2    b   2 1  PQ     PQR
#> 3 175.11895  y1    y   1 1   R     PQR
#> 4 303.17753  y2    y   2 1  QR     PQR
#> 5 157.10839 y1_   y_   1 1   R     PQR
#> 6 285.16697 y2_   y_   2 1  QR     PQR
#> 7 286.15098 y2*   y*   2 1  QR     PQR
calculateFragments("PQR",
                   neutralLoss=defaultNeutralLoss(disableWaterLoss="Cterm"))
#> Fixed modifications used: C=57.02146
#> Variable modifications used: None
#>          mz ion type pos z seq peptide
#> 1  98.06004  b1    b   1 1   P     PQR
#> 2 226.11862  b2    b   2 1  PQ     PQR
#> 3 175.11895  y1    y   1 1   R     PQR
#> 4 303.17753  y2    y   2 1  QR     PQR
#> 5 286.15098 y2*   y*   2 1  QR     PQR
calculateFragments("PQR",
                   neutralLoss=defaultNeutralLoss(disableAmmoniaLoss="Q"))
#> Fixed modifications used: C=57.02146
#> Variable modifications used: None
#>          mz ion type pos z seq peptide
#> 1  98.06004  b1    b   1 1   P     PQR
#> 2 226.11862  b2    b   2 1  PQ     PQR
#> 3 175.11895  y1    y   1 1   R     PQR
#> 4 303.17753  y2    y   2 1  QR     PQR
#> 5 157.10839 y1_   y_   1 1   R     PQR
#> 6 285.16697 y2_   y_   2 1  QR     PQR

## disable neutral loss completely
calculateFragments("PQR", neutralLoss=NULL)
#> Fixed modifications used: C=57.02146
#> Variable modifications used: None
#>          mz ion type pos z seq peptide
#> 1  98.06004  b1    b   1 1   P     PQR
#> 2 226.11862  b2    b   2 1  PQ     PQR
#> 3 175.11895  y1    y   1 1   R     PQR
#> 4 303.17753  y2    y   2 1  QR     PQR